requirements.txt
geopandas==0.13.0
fiona==1.9.4
shapely==2.0.1
pyproj==3.6.0
google-cloud-storage==2.11.0
googlemaps==4.10.0
pandas==2.2.3
matplotlib==3.8.0
python-dotenv==1.0.0
